liga_mx_bot/web_nueva/requirements.txt
flask==2.2.3
prometheus-client==0.16.0
flask-prometheus-metrics==1.0.0
python-dotenv==1.0.0
requests==2.28.2
pytz==2023.3
pyyaml==6.0
structlog==23.1.0
